GREST-78 and GREST-80.
Refactored BasicDomainClassMarshaller and created unit tests for it.
BasicDomainClassMarshaller can now make decisions on what fields to include (and how to name them) on a per-instance basis. Making the decisions on a per-instance basis allows for creation of sublclasses that use annotations on domain classes to decide what fields to include, etc.
It can either include all persistent fields not excluded, or it can only include fields explicity included.

In general, it is expected that this class will be overriden by classes that will provide implementations of the methods via configuration.