Add MapTransformer and related tests.
MapTransformer is a utility class to support
transformation and 'flattening' of maps
to faciliate the use of grails data binding to
domain and pogo classes.  It can be used
to rename keys, add required default values,
modify values (such as turning a short object
representation into a bare id) and 'flatten'
maps into the index and dot notation used by
grails data binding.