Change content-type behavior for non-json/xml.
Previously, if the media-type selected for a response
representation did not end with xml or json, the
Content-Type header was set to application/text.

Now, if the media type ends in json or xml, the
Content-Type header is set to application/json or
application/xml.  Otherwise, it is set to the
value of the selected media type (unless contentType
is explicitly set on the representation).

This behavior is more in line with what you would
expect when dealing with text/calendar or application/pdf
representations.