Update param and content id checking.
Fix comparison between id field in content and params
to convert the content field to a string first.

Add ability to configure a resource to ignore the id
field in content.

By default, the update and delete operations will check
incoming content for an id field, and if found, verify
that the field (converted to a string) matches the
id parameter extracted from the url.  This behavior
can be disabled by setting idMatchEnforced to false
on the resource configuration.